Update the table column to decimal

+++ src/Tickets/Order_Modifiers/Custom_Tables/Order_Modifiers.php
@@ -69,7 +69,7 @@ class Order_Modifiers extends Abstract_Custom_Table {
 				`id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
 				`modifier_type` VARCHAR(255) NOT NULL,
 				`sub_type` VARCHAR(255) NOT NULL,
-				`raw_amount` INT NOT NULL,
+				`raw_amount` DECIMAL(18,6) NOT NULL,
 				`slug` VARCHAR(255) NOT NULL,
 				`display_name` VARCHAR(255) NOT NULL,
 				`status` VARCHAR(20) NOT NULL DEFAULT 'draft',
